Fix family validation
