+++ scripts/ingest_dfc/utils.py
@@ -1,11 +1,13 @@
 import enum
+from dataclasses import fields
 from datetime import datetime
-from typing import TypeVar
+from typing import Any, Sequence, TypeVar
 
+from pydantic import ConfigDict, Extra
+from pydantic.dataclasses import dataclass
 from sqlalchemy.orm import Session
 
 from app.db.session import Base
-
 
 class PublicationDateAccuracy(enum.IntEnum):
     """
@@ -24,6 +26,131 @@ class PublicationDateAccuracy(enum.IntEnum):
 """An undefined datetime"""
 # FIXME: We may choose to set this to `None` instead & make the date field nullable
 UNDEFINED_DATA_TIME = datetime(1900, 1, 1, 0, 0, 0, PublicationDateAccuracy.NOT_DEFINED)
+
+
+REQUIRED_COLUMNS = [
+    "ID",
+    "Document ID",
+    "CCLW Description",
+    "Part of collection?",
+    "Create new family/ies?",
+    "Collection ID",
+    "Collection name",
+    "Collection summary",
+    "Document title",
+    "Family name",
+    "Family summary",
+    "Family ID",
+    "Document role",
+    "Applies to ID",
+    "Geography ISO",
+    "Documents",
+    "Category",
+    "Events",
+    "Sectors",
+    "Instruments",
+    "Frameworks",
+    "Responses",
+    "Natural Hazards",
+    "Document Type",
+    "Year",
+    "Language",
+    "Keywords",
+    "Geography",
+    "Parent Legislation",
+    "Comment",
+    "CPR Document ID",
+    "CPR Family ID",
+    "CPR Collection ID",
+    "CPR Family Slug",
+    "CPR Document Slug",
+]
+
+VALID_COLUMN_NAMES = set(REQUIRED_COLUMNS)
+
+
+def validate_csv_columns(column_names: Sequence[str]) -> bool:
+    return VALID_COLUMN_NAMES.issubset(set(column_names))
+
+
+@dataclass(config=ConfigDict(validate_assignment=True, extra=Extra.forbid))
+class DfcRow:
+    row_number: int
+    id: str
+    document_id: str
+    cclw_description: str
+    part_of_collection: str
+    create_new_families: str
+    collection_id: str
+    collection_name: str
+    collection_summary: str
+    document_title: str
+    family_name: str
+    family_summary: str
+    family_id: str
+    document_role: str
+    applies_to_id: str
+    geography_iso: str
+    documents: str
+    category: str
+    events: list[str]
+    sectors: list[str]
+    instruments: list[str]
+    frameworks: list[str]
+    responses: list[str]
+    natural_hazards: list[str]
+    keywords: list[str]
+    document_type: str
+    year: int
+    language: str
+    geography: str
+    parent_legislation: str
+    comment: str
+    cpr_document_id: str
+    cpr_family_id: str
+    cpr_collection_id: str
+    cpr_family_slug: str
+    cpr_document_slug: str
+
+    @classmethod
+    def from_row(cls, row_number: int, data: dict[str, str]):
+        """Parse a row from a CSV into the DfcRow type"""
+        field_info = cls.field_info()
+        return cls(
+            row_number=row_number,
+            **{
+                cls._key(k): cls._parse_str(cls._key(k), v, field_info)
+                for (k, v) in data.items()
+            },
+        )
+
+    @classmethod
+    def field_info(cls) -> dict[str, type]:
+        return {field.name: field.type for field in fields(cls)}
+
+    @classmethod
+    def _parse_str(cls, key: str, value: str, field_info: dict[str, type]) -> Any:
+        if key not in field_info:
+            # Let pydantic deal with unexpected fields
+            return value
+
+        if field_info[key] == list[str]:
+            return [e.strip() for e in value.split(";")]
+
+        if field_info[key] == int:
+            return int(value) if value else 0
+
+        if field_info[key] == str:
+            if (na := str(value).lower()) == "n/a":
+                return na
+            else:
+                return value
+
+        raise Exception(f"Unhandled type '{cls.field_info()[key]}' in row parsing")
+
+    @staticmethod
+    def _key(key: str) -> str:
+        return key.lower().replace(" ", "_").replace("?", "").replace("y/", "")
 
 
 DbModel = TypeVar('DbModel', bound=Base)
